Unify Provider and Consumer library
This commit relocates the Provider and Consumer charm libraries
into a single module which is the Prometheus charm library. There
are no changes to the functionality of existing objects here, and
this commit only involes a relocation of code.
