+++ tests/test_consumer.py
